Some work on view classes
Two callbacks removed from model.Player class, these were originally to provide an easy way of listening for all frame/ball score changes for a particular player, but they probably aren't going to be needed.
Order of callback arguments changed, to simplify callbacks in the view classes. Putting the changed data as the first argument means that setSomething(x) functions can be called directly and the other arguments will be ignored. Otherwise an extra function is needed to pass the correct argument to setSomething(x).